perf: improve performance of Overwrite and Merge
Previously Overwrite was written in terms of Merge, and Merge was not
particularly performant (many iterations over object keys). With TS 2.8,
Overwrite became much easier to write with a single iteration over the
left object's keys, and no iterations over the right object's keys. This
should lead to dramatic performance gains for Overwrite.

To take advantage of these performance benefits, Merge is now written in
terms of an Overwrite operation; instead of the other way around. This
should result in a small improvement in Merge's performance.