perf: greatly simplify `IsAny` type
Thanks to @krisdages for pointing out this simplification! This PR makes
the `IsAny` type much more simple. Previously we were exhaustively
checking that the given type, `T`, was _not_ any other type. This is
pretty prone to mistake and is not very sustainable (a new fundamental
type could be introduced).