docs: add ability to filter regression tests
This commit will filter out regression tests that are stored in the
unit test files. I started to refactor tests into a `unit` and
`regression` subfolder, but that seemed like a lot of refactor for
1) a bug fix PR
2) a single regression test.

I'll consider making the separation more clear in the future. For now,
I add the ability to hide tests from the documentation generator. This
script needs to be cleaned up, so I'll look into better ways to handle
this in future PRs.