test: add regression/failing test for KeysByType
The regression test makes sure that this PR actually solves the problem
that was brought up in issue #106. It also adds a failing test case for
a bug that I noticed while making the fix. Since my ad hoc testing
solution obviously does not allow for known failures (which I guess is,
on its own, a known failure..), I annotated this with a comment. There
is obviously a lot of work to be done on the testing side of this
library.