test: drop older versions of typescript
The new `IsNever` only works with 3.5, so all older versions need to be
dropped. This will bump the major version of `simplytyped` to avoid any
breaking changes.

I think it is worth it to drop older versions of the language to keep
this package from rotting too severely, as long as we are careful with
updating the major version whenever we change the supported TS versions.

BREAKING CHANGE: no longer supporting typescript < 3.5